payments/src/events/listener/__test__/order-created-listener.test.ts
import { OrderCreatedEvent, OrderStatus } from '@mrltickets/common';
import mongoose from 'mongoose';
import { Message } from 'node-nats-streaming';
import { Order } from '../../../models/order';
import { natsWrapper } from '../../../nats-wrapper';
import { OrderCreatedListener } from '../order-created-listener';

const setup = async () => {
  //create  an instance of the  listener
  const listener = new OrderCreatedListener(natsWrapper.client);

  //create a fake data object
  const data: OrderCreatedEvent['data'] = {
    id: new mongoose.Types.ObjectId().toHexString(),
    version: 0,
    expiresAt: new Date().toISOString(),
    userId: new mongoose.Types.ObjectId().toHexString(),
    status: OrderStatus.Created,
    ticket: {
      id: new mongoose.Types.ObjectId().toHexString(),
      price: 19,
    },
    quantity: 10,
  };

  //create a  fake  message object
  //@ts-ignore
  const msg: Message = {
    ack: jest.fn(),
  };

  //return  all the stuffs
  return { listener, data, msg };
};

it('replicates the order  info', async () => {
  const { listener, data, msg } = await setup();

  await listener.onMessage(data, msg);

  const order = await Order.findById(data.id);

  expect(order).toBeDefined();
  expect(order!.price).toEqual(data.ticket.price * data.quantity);
});

it('acks  the message', async () => {
  const { listener, data, msg } = await setup();

  await listener.onMessage(data, msg);

  expect(msg.ack).toHaveBeenCalled();
});
